requirements.txt
fastapi
psycopg2-binary
python-decouple
asyncpg
sqlalchemy
alembic
passlib
PyJWT
pydantic
pydantic-settings
uvicorn
transformers
torch
python-docx
PyPDF2
